+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Fix return type of `sortBy` and `sortByDescending` methods
 
 ## 8.1.0 - 2023-05-19
 - Add `Immutable\Generic\IList::choose` method